Tests/LibWeb/Text/input/css/native-var-substitution.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    test(() => {
        internals.updateStyle();
        internals.resetStyleFfiCounters();

        const element = document.createElement("div");
        element.style.cssText = `
            --base: 5px;
            --nested: calc(var(--base) * 2);
            --cycle-a: var(--cycle-b);
            --cycle-b: var(--cycle-a);
            --cycle-fallback-a: var(--cycle-fallback-b, green);
            --cycle-fallback-b: var(--cycle-fallback-a);
            --cycle-dependent: var(--cycle-fallback-a, blue);
            --\\78: 7px;
            width: var(--nested);
            height: var(--\\78);
            margin-left: var(--cycle-a, 3px);
            color: var(--cycle-fallback-a, red);
            background-color: var(--cycle-dependent, black);
        `;
        document.body.appendChild(element);

        const computed = getComputedStyle(element);
        println(`width: ${computed.width}`);
        println(`height: ${computed.height}`);
        println(`cycle fallback: ${computed.marginLeft}`);
        println(`fallback inside cycle ignored: ${computed.color}`);
        println(`fallback outside cycle used: ${computed.backgroundColor}`);

        const nativeCounters = internals.styleFfiCounters();
        println(`native parse callback used: ${nativeCounters.cascadeParseSubstitutedCallbacks >= 3}`);
        println(`C++ resolver unused: ${nativeCounters.cascadeResolveUnresolvedCallbacks === 0}`);

        CSS.registerProperty({
            name: "--registered-native-fallback",
            syntax: "<length>",
            inherits: false,
            initialValue: "1px",
        });
        CSS.registerProperty({
            name: "--registered-universal-native",
            syntax: "*",
            inherits: false,
            initialValue: "13px",
        });
        internals.resetStyleFfiCounters();

        const registered = document.createElement("div");
        registered.style.cssText = `
            --registered-native-fallback: 9px;
            --unregistered-native-after-registration: 11px;
            width: var(--registered-native-fallback);
            height: var(--unregistered-native-after-registration);
            margin-left: var(--registered-universal-native);
        `;
        document.body.appendChild(registered);
        println(`registered width: ${getComputedStyle(registered).width}`);
        println(`unregistered height: ${getComputedStyle(registered).height}`);
        println(`registered universal initial: ${getComputedStyle(registered).marginLeft}`);

        const registeredCounters = internals.styleFfiCounters();
        println(`registered value used C++ resolver: ${registeredCounters.cascadeResolveUnresolvedCallbacks >= 1}`);
        println(`unregistered value used native parser: ${registeredCounters.cascadeParseSubstitutedCallbacks >= 1}`);
    });
</script>
