Tests/LibWeb/Text/expected/css/native-var-substitution.txt

width: 10px
height: 7px
cycle fallback: 3px
fallback inside cycle ignored: rgb(255, 0, 0)
fallback outside cycle used: rgb(0, 0, 255)
native parse callback used: true
C++ resolver unused: true
registered width: 9px
unregistered height: 11px
registered universal initial: 13px
registered value used C++ resolver: true
unregistered value used native parser: true
